fix: Use HMR API for dev server communication
